refactored command to commands
