+++ commands/apikeys/delete.go
