+++ commands/nodepool/delete.go
