+++ commands/nodepool/describe.go
