+++ commands/nodepool/list.go
